requirements.txt
Flask
Flask-SQLAlchemy
spaCy
python-dateutil
dateparser

